refactored props passed to MusicContainer
